+++ conman/routes/handlers.py
@@ -1,3 +1,4 @@
+from django.core import checks
 from django.urls import resolve, Resolver404
 
 
@@ -7,14 +8,19 @@ class BaseHandler:
 
     Subclasses should define `handle`.
     """
-    @classmethod
-    def path(cls):
-        """Get dotted-path of this class."""
-        return '{}.{}'.format(cls.__module__, cls.__name__)
-
     def __init__(self, route):
         """Store the Route so that we know what we're handling."""
         self.route = route
+
+    @classmethod
+    def check(cls, route):
+        """
+        Model-level checks for a Route type delegated to its Handler.
+
+        Custom classes can override this to perform checks on the Routes they
+        are attached to.
+        """
+        return []
 
     def handle(self, request, path):
         """Raise an error if a subclass calls handle without defining how.."""
@@ -32,6 +38,21 @@ class URLConfHandler(BaseHandler):
     Views referenced in the `urlconf` will receive `route`, as well as the args
     and kwargs they would expect given their urlpattern.
     """
+    @classmethod
+    def check(cls, route):
+        """Ensure route has a sensible urlconf attribute."""
+        if not hasattr(route, 'urlconf'):
+            return [checks.Error(
+                '{} must have a `urlconf` attribute.'.format(route.__name__),
+                hint=(
+                    'The urlconf must be a dotted path. ' +
+                    'This is a requirement of {}.'.format(cls.__name__)
+                ),
+                obj=route,
+            )]
+
+        return []
+
     def handle(self, request, path):
         """
         Resolve `path` to a view, and get it to handle the `request`.
@@ -53,6 +74,18 @@ class RouteViewHandler(BaseHandler):
 
     Views will receive `route` as a keyword arg.
     """
+    @classmethod
+    def check(cls, route):
+        """Ensure route has a sensible view attribute."""
+        if not hasattr(route, 'view'):
+            return [checks.Error(
+                '{} must have a `view` attribute.'.format(route.__name__),
+                hint='This is a requirement of {}.'.format(cls.__name__),
+                obj=route,
+            )]
+
+        return []
+
     def handle(self, request, path):
         """
         Handle the request using the `view` attribute of the associated route.
@@ -66,5 +99,6 @@ class RouteViewHandler(BaseHandler):
         """
         if path != '/':
             raise Resolver404
-        view = self.route.view
-        return view(request, route=self.route)
+        # We use `type` here to ensure that we don't access the view as a bound
+        # method of the Route, but instead get the view as a function.
+        return type(self.route).view(request, route=self.route)